refactor(services): split context_loader into a package (Phase 3)
context_loader.py had grown to 1,316 lines — token estimation, similarity
scoring, condensation policy, completed-roadmap parsing, and the main
ContextLoader orchestrator were all piled in one file. Split into a
package where the pieces stand on their own:

  services/context_loader/
    __init__.py     — re-exports the public API (back-compat)
    utils.py        — estimate_tokens, truncate_content, extract_keywords,
                      compute_similarity_scores + optional-dep probes (229 LOC)
    condenser.py    — ContextCondenser class, guidance + priority truncation (102 LOC)
    completed.py    — parse_completed_roadmap, format_completed_for_context (92 LOC)
    loader.py       — ContextLoader orchestrator (601 LOC)

All existing imports continue working unchanged because the package
__init__.py re-exports every public name. 56 unit tests for context_loader
still pass without changes. No behavior or API changes.

Scope notes — deliberately deferred from this phase:

- template_manager.py (926 LOC) is a single cohesive TemplateManager class;
  splitting it into mixins/helpers would worsen cohesion without clear gain.
  Addressing its complexity needs a concept-level refactor, not a mechanical
  one.
- github_* consolidation (as originally written in the plan) would merge
  github_linker.py into the already-legacy github_service.py (noted in its
  module docstring as back-compat shim for the provider abstraction).
  That would grow the god-object, not shrink it. Correct follow-up is to
  migrate callers off github_service.py toward services/providers/ and then
  delete the shim — out of scope for this PR.

Verification: 1,735 unit/contract/integration tests pass. ruff clean.
CLI smoke tests (`doit --help`, `doit status`) unchanged.
